Add art patch
Android runtime has an exception where 64bit dex2oat will not perform to its full potential if you are on a kernel older than 3.12 to avoid a kernel bug.

Our kernel includes this fix in the samsung src already, so we are just leaving performance on the table due to an arbitrary number limit. Get rid of this check to use full performance.